o created new project to split integration with SpringBoot
