+++ juneau-microservice/juneau-microservice-server-springboot/src/main/java/org/apache/juneau/microservice/springboot/annotations/EnabledJuneauIntegration.java
@@ -10,43 +10,21 @@
 // * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the        *
 // * specific language governing permissions and limitations under the License.                                              *
 // ***************************************************************************************************************************
-package org.apache.juneau.examples.rest.springboot;
+package org.apache.juneau.microservice.springboot.annotations;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.*;
-import org.springframework.context.*;
-import org.springframework.context.annotation.*;
-import org.springframework.stereotype.*;
+import org.apache.juneau.microservice.BasicRestServletJenaGroup;
+import org.apache.juneau.microservice.springboot.config.AppConfiguration;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
+import org.springframework.stereotype.Component;
 
-/**
- * Entry point for Examples REST application when deployed as a Spring Boot application.
- */
-@SpringBootApplication
-@Controller
-@Import({AppConfiguration.class, AppServletConfiguration.class})
-public class App {
+import java.lang.annotation.*;
 
-	public static int counter = 0;
-	private static volatile ConfigurableApplicationContext context;
-
-	public static void main(String[] args) {
-		if (System.getProperty("juneau.configFile") == null)
-			System.setProperty("juneau.configFile", "examples.cfg");
-		try {
-			context = SpringApplication.run(App.class, args);
-			if (context == null)
-				System.exit(2); // Probably port in use?
-			AppConfiguration.setAppContext(context);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	public static void start() {
-		main(new String[0]);
-	}
-
-	public static void stop() {
-		context.stop();
-	}
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Inherited
+@Component
+@ImportAutoConfiguration(classes = AppConfiguration.class)
+public @interface EnabledJuneauIntegration {
+    Class<? extends BasicRestServletJenaGroup> rootResources();
 }
